pyproject.toml
[tool.poetry]
name = "luizalabs-interview"
version = "0.1.0"
description = ""
authors = ["Fernando Toledo <fernandohtr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
django-redis = "^5.4.0"
psycopg2 = "^2.9.9"
gunicorn = "^23.0.0"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pytest-django = "^4.9.0"
pytest-factoryboy = "^2.7.0"
pytest-cov = "^5.0.0"
ipython = "^8.28.0"
responses = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
